requirements.txt
requests
pymongo
python-dotenv
duckdb
pandas
streamlit
plotly
